Add files to run a dummy experiment
The dummy experiment is refactored to mimic the way other experiments
are run: using a bash script in project directory to execute python
files located in Experiments/dummy directory for training and evaluation.

The dummy experiment trains the network on a synthetic dataset for three
tasks -- two regression tasks and one classification task. The trained
network is saved in the Experiments directory. The saved model can be
loaded and evaluated on a synthetic test dataset.